+++ Labs/Lab04/ConsoleApp1/Model/EmployeeTerritory.cs
@@ -1,0 +1,13 @@
+﻿namespace ConsoleApp1.Model;
+
+public class EmployeeTerritory
+{
+    public string EmployeeId { get; set; }
+    public string TerritoryId { get; set; }
+
+    public EmployeeTerritory(string employeeId, string territoryId)
+    {
+        EmployeeId = employeeId;
+        TerritoryId = territoryId;
+    }
+}
