+++ Labs/Lab04/ConsoleApp1/Model/Territory.cs
@@ -1,0 +1,15 @@
+﻿namespace ConsoleApp1.Model;
+
+public class Territory
+{
+    public string TerritoryId { get; set; }
+    public string TerritoryDescription { get; set; }
+    public string RegionId { get; set; }
+
+    public Territory(string territoryId, string description, string regionId)
+    {
+        TerritoryId = territoryId;
+        TerritoryDescription = description;
+        RegionId = regionId;
+    }
+}
